Fix flux calculation, add Kerr flux calculation
Multiple things were wrong during flux calculation:
- The log should be the natural log, not base10 log
- Luminet mistakenly reports the factor before the log to be sqrt(3)/3,
but it should be sqrt(3)/2

I verified the math from Page and Thorne 1974 for this. Solving Equation
15n for a Swarzschild metric (where a=0) eventually lands on Luminet's
equation 15, except for the aforementioned factor.

While I was verifying the math from Page 1974 anyways (which is Kerr), I
implemented the routines needed to calculate innermost stable orbits and
flux for Kerr in general. This will be of use when I eventually extend
this repo to include Kerr metrics.